feat : create a separate util folder for helper functions and implement year utils
