+++ src/utils/__init__.py
@@ -1,0 +1,6 @@
+from .year_data_utils import scrape_years_metadata, load_scraped_years
+
+__all__ = [
+    "scrape_years_metadata",
+    "load_scraped_years"
+]
